♻️ Refactor : 툴바 모드에 따라서 분기처리 리팩토링
